requirements.txt
langgraph>=0.2
langchain>=0.3
langchain-core>=0.3
langchain-community>=0.3
langchain-ollama>=0.2
llama-index-core>=0.11
llama-index-vector-stores-qdrant>=0.1
llama-index-embeddings-huggingface>=0.3
qdrant-client>=1.12
langchain-tongyi>=0.2
dashscope>=1.20
pytest>=8.0
fastapi>=0.115
uvicorn>=0.30
